test/reftest/web-platform-tests/dom/nodes/Element-childElementCount-dynamic-add-xhtml.xhtml

<?xml version="1.0" encoding="UTF-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Dynamic Adding of Elements</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<div>Test of Dynamic Adding of Elements</div>
<div id="log"></div>
<p id="parentEl">The result of this test is
<span id="first_element_child" style="font-weight:bold;">logged above.</span></p>
<script>
test(function() {
  var parentEl = document.getElementById("parentEl");
  var newChild = document.createElement("span");
  parentEl.appendChild(newChild);
  assert_equals(parentEl.childElementCount, 2)
})
</script>
</body>
</html>
